# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "VocaDbWeb",
    "version": "1.0"
  },
  "paths": {
...
    "/api/comments/{entryType}-comments/{commentId}": {
      "delete": {
        "tags": [
          "CommentApi"
        ],
        "parameters": [
          {
            "name": "entryType",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntryType"
            }
          },
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "CommentApi"
        ],
        "parameters": [
          {
            "name": "entryType",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntryType"
            }
          },
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CommentForApiContract"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentForApiContract"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentForApiContract"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CommentForApiContract"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/comments/{entryType}-comments": {
      "get": {
        "tags": [
          "CommentApi"
        ],
        "parameters": [
          {
            "name": "entryType",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntryType"
            }
          },
          {
            "name": "entryId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CommentForApiContractPartialFindResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentForApiContractPartialFindResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentForApiContractPartialFindResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CommentApi"
        ],
        "parameters": [
          {
            "name": "entryType",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntryType"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CommentForApiContract"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentForApiContract"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentForApiContract"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CommentForApiContract"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CommentForApiContract"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentForApiContract"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentForApiContract"
                }
              }
            }
          }
        }
      }
    },
    "/api/comments": {
      "get": {
        "tags": [
          "CommentApi"
        ],
        "parameters": [
          {
            "name": "before",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "since",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "entryType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EntryType"
            }
          },
          {
            "name": "maxResults",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "getTotalCount",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "fields",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/CommentOptionalFields"
            }
          },
          {
            "name": "entryFields",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EntryOptionalFields"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ContentLanguagePreference"
            }
          },
          {
            "name": "sortRule",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/CommentSortRule"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CommentForApiContractPartialFindResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentForApiContractPartialFindResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentForApiContractPartialFindResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtistType": {
        "enum": [
          "Unknown",
          "Circle",
          "Label",
          "Producer",
          "Animator",
          "Illustrator",
          "Lyricist",
          "Vocaloid",
          "UTAU",
          "CeVIO",
          "OtherVoiceSynthesizer",
          "OtherVocalist",
          "OtherGroup",
          "OtherIndividual",
          "Utaite",
          "Band",
          "Vocalist",
          "Character",
          "SynthesizerV",
          "CoverArtist",
          "NEUTRINO",
          "VoiSona",
          "NewType",
          "Voiceroid",
          "Instrumentalist",
          "Designer"
        ],
        "type": "string"
      },
      "CommentForApiContract": {
        "type": "object",
        "properties": {
          "author": {
            "$ref": "#/components/schemas/UserForApiContract"
          },
          "authorName": {
            "type": "string",
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "entry": {
            "$ref": "#/components/schemas/EntryForApiContract"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CommentForApiContractPartialFindResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentForApiContract"
            },
            "nullable": true
          },
          "term": {
            "type": "string",
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CommentOptionalFields": {
        "enum": [
          "None",
          "Entry"
        ],
        "type": "string"
      },
      "CommentSortRule": {
        "enum": [
          "CreateDateDescending",
          "CreateDate"
        ],
        "type": "string"
      },
      "ContentLanguagePreference": {
        "enum": [
          "Default",
          "Japanese",
          "Romaji",
          "English"
        ],
        "type": "string"
      },
      "ContentLanguageSelection": {
        "enum": [
          "Unspecified",
          "Japanese",
          "Romaji",
          "English"
        ],
        "type": "string"
      },
...
      "DiscType": {
        "enum": [
          "Unknown",
          "Album",
          "Single",
          "EP",
          "SplitAlbum",
          "Compilation",
          "Video",
          "Artbook",
          "Game",
          "Fanmade",
          "Instrumental",
          "Other"
        ],
        "type": "string"
      },
...
      "EntryForApiContract": {
        "type": "object",
        "properties": {
          "activityDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "additionalNames": {
            "type": "string",
            "nullable": true
          },
          "artistString": {
            "type": "string",
            "nullable": true
          },
          "artistType": {
            "$ref": "#/components/schemas/ArtistType"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "defaultName": {
            "type": "string",
            "nullable": true
          },
          "defaultNameLanguage": {
            "$ref": "#/components/schemas/ContentLanguageSelection"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "discType": {
            "$ref": "#/components/schemas/DiscType"
          },
          "entryType": {
            "$ref": "#/components/schemas/EntryType"
          },
          "eventCategory": {
            "$ref": "#/components/schemas/EventCategory"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "mainPicture": {
            "$ref": "#/components/schemas/EntryThumbForApiContract"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "names": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LocalizedStringContract"
            },
            "nullable": true
          },
          "pvs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PVContract"
            },
            "nullable": true
          },
          "songListFeaturedCategory": {
            "$ref": "#/components/schemas/SongListFeaturedCategory"
          },
          "songType": {
            "$ref": "#/components/schemas/SongType"
          },
          "status": {
            "$ref": "#/components/schemas/EntryStatus"
          },
          "releaseEventSeriesName": {
            "type": "string",
            "nullable": true
          },
          "tagCategoryName": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagUsageForApiContract"
            },
            "nullable": true
          },
          "urlSlug": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "webLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArchivedWebLinkContract"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EntryOptionalFields": {
        "enum": [
          "None",
          "AdditionalNames",
          "Description",
          "MainPicture",
          "Names",
          "PVs",
          "Tags",
          "WebLinks"
        ],
        "type": "string"
      },
...
      "EntryThumbForApiContract": {
        "type": "object",
        "properties": {
          "mime": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "urlOriginal": {
            "type": "string",
            "nullable": true
          },
          "urlSmallThumb": {
            "type": "string",
            "nullable": true
          },
          "urlThumb": {
            "type": "string",
            "nullable": true
          },
          "urlTinyThumb": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EntryType": {
        "enum": [
          "Undefined",
          "Album",
          "Artist",
          "DiscussionTopic",
          "PV",
          "ReleaseEvent",
          "ReleaseEventSeries",
          "Song",
          "SongList",
          "Tag",
          "User",
          "Venue"
        ],
        "type": "string"
      },
...
      "EventCategory": {
        "enum": [
          "Unspecified",
          "AlbumRelease",
          "Anniversary",
          "Club",
          "Concert",
          "Contest",
          "Convention",
          "Other",
          "Festival"
        ],
        "type": "string"
      },
...
      "LocalizedStringContract": {
        "type": "object",
        "properties": {
          "language": {
            "$ref": "#/components/schemas/ContentLanguageSelection"
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OldUsernameContract": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "oldName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UserForApiContract": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean"
          },
          "groupId": {
            "$ref": "#/components/schemas/UserGroupId"
          },
          "knownLanguages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserKnownLanguageContract"
            },
            "nullable": true
          },
          "mainPicture": {
            "$ref": "#/components/schemas/EntryThumbForApiContract"
          },
          "memberSince": {
            "type": "string",
            "format": "date-time"
          },
          "oldUsernames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OldUsernameContract"
            },
            "nullable": true
          },
          "verifiedArtist": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UserGroupId": {
        "enum": [
          "Nothing",
          "Limited",
          "Regular",
          "Trusted",
          "Moderator",
          "Admin"
        ],
        "type": "string"
      },
...
      "UserKnownLanguageContract": {
        "type": "object",
        "properties": {
          "cultureCode": {
            "type": "string",
            "nullable": true
          },
          "proficiency": {
            "$ref": "#/components/schemas/UserLanguageProficiency"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}